# docs/openapi/ioas-openapi.yaml
openapi: 3.0.3
info:
  title: Tree Education IOAS API
  version: v1
  description: 正式后端接口契约，覆盖认证、运营人员、媒体主题包、素材、资源中心、线索、任务、报表、回收站与审计。
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/media/content/packages/{id}:
    get:
      tags: [Content Package]
      summary: 主题包详情
      parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
      responses: {'200': {description: 主题包详情}, '404': {description: 主题包不存在}}
    put:
      tags: [Content Package]
      summary: 编辑主题包
      parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
      responses: {'200': {description: 已更新主题包}}
    delete:
      tags: [Content Package]
      summary: 删除空主题包
      parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
      responses: {'200': {description: 删除成功}, '400': {description: 主题包仍包含文件}}
  /api/v1/media/assets/packages/{packageId}:
    post:
      tags: [Asset File]
      summary: 上传文件到主题包；文件不可脱离主题包存在
      parameters: [{name: packageId, in: path, required: true, schema: {type: integer, format: int64}}, {name: type, in: query, required: true, schema: {$ref: '#/components/schemas/AssetFileType'}}]
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, properties: {file: {type: string, format: binary}, description: {type: string}}}}}}
      responses: {'200': {description: 已上传素材}}
...
  /api/v1/leads/{id}:
    get:
      tags: [Lead]
      summary: 线索详情
      parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
      responses: {'200': {description: 线索详情}}
  /api/v1/leads/{id}/status:
    patch:
      tags: [Lead]
      summary: 更新线索状态
      parameters: [{name: id, in: path, required: true, schema: {type: integer, format: int64}}]
      responses: {'200': {description: 已更新线索}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
    AssetFileType: {type: string, enum: [script, video, image]}
...