# src/main/java/com/treeeducation/ioas/lead/LeadController.java
package com.treeeducation.ioas.lead; import com.treeeducation.ioas.auth.UserPrincipal;import com.treeeducation.ioas.common.*;import io.swagger.v3.oas.annotations.*;import io.swagger.v3.oas.annotations.tags.Tag;import jakarta.validation.Valid;import org.springframework.data.domain.*;import org.springframework.security.core.annotation.AuthenticationPrincipal;import org.springframework.web.bind.annotation.*; /** Lead APIs. */ @RestController @RequestMapping("/api/v1/leads") @Tag(name="Lead",description="线索列表、详情、创建与状态更新") public class LeadController{private final LeadService service; public LeadController(LeadService service){this.service=service;} @PostMapping @Operation(summary="基于主题包和素材创建线索") public ApiResponse<LeadDtos.Response> create(@Valid @RequestBody LeadDtos.CreateRequest r,@AuthenticationPrincipal UserPrincipal p){return ApiResponse.ok(LeadDtos.of(service.create(r,p)));} @GetMapping @Operation(summary="线索列表") public ApiResponse<PageResponse<LeadDtos.Response>> list(@RequestParam(required=false)Long packageId,@RequestParam(defaultValue="0")int page,@RequestParam(defaultValue="20")int size){return ApiResponse.ok(PageResponse.of(service.list(packageId,PageRequest.of(page,size))));} @GetMapping("/{id}") @Operation(summary="线索详情") public ApiResponse<LeadDtos.Response> get(@PathVariable Long id){return ApiResponse.ok(LeadDtos.of(service.get(id)));} @PatchMapping("/{id}/status") @Operation(summary="更新线索状态") public ApiResponse<LeadDtos.Response> status(@PathVariable Long id,@Valid @RequestBody LeadDtos.StatusRequest r){return ApiResponse.ok(LeadDtos.of(service.updateStatus(id,r.status())));} }