# src/main/java/com/treeeducation/ioas/media/contentpackage/ContentPackageController.java
package com.treeeducation.ioas.media.contentpackage;
import com.treeeducation.ioas.auth.UserPrincipal;import com.treeeducation.ioas.common.*;import io.swagger.v3.oas.annotations.*;import io.swagger.v3.oas.annotations.tags.Tag;import jakarta.validation.Valid;import org.springframework.data.domain.*;import org.springframework.security.core.annotation.AuthenticationPrincipal;import org.springframework.web.bind.annotation.*;
/** REST API for content packages. */
@RestController @RequestMapping("/api/v1/media/content/packages") @Tag(name="Content Package",description="媒体主题包管理") public class ContentPackageController{ private final ContentPackageService service; public ContentPackageController(ContentPackageService service){this.service=service;} @PostMapping @Operation(summary="新建主题包") public ApiResponse<ContentPackageDtos.Response> create(@Valid @RequestBody ContentPackageDtos.UpsertRequest r,@AuthenticationPrincipal UserPrincipal p){return ApiResponse.ok(ContentPackageDtos.of(service.create(r,p)));} @GetMapping @Operation(summary="主题包列表") public ApiResponse<PageResponse<ContentPackageDtos.Response>> list(@RequestParam(defaultValue="0")int page,@RequestParam(defaultValue="20")int size){return ApiResponse.ok(PageResponse.of(service.list(PageRequest.of(page,size,Sort.by(Sort.Direction.DESC,"createdAt")))));} @GetMapping("/{id}") @Operation(summary="主题包详情") public ApiResponse<ContentPackageDtos.Response> get(@PathVariable Long id){return ApiResponse.ok(ContentPackageDtos.of(service.get(id)));} @PutMapping("/{id}") @Operation(summary="编辑主题包") public ApiResponse<ContentPackageDtos.Response> update(@PathVariable Long id,@Valid @RequestBody ContentPackageDtos.UpsertRequest r){return ApiResponse.ok(ContentPackageDtos.of(service.update(id,r)));} @DeleteMapping("/{id}") @Operation(summary="删除空主题包") public ApiResponse<Void> delete(@PathVariable Long id){service.delete(id);return ApiResponse.ok();}}